fix: resolve Sonar S5276 and S1820 issues in PR 50
S5276 — add explicit (float) casts on TILE_SIZE and offset_y in ghost.c
and fruit.c popup draw calculations to eliminate implicit int→float
conversions.

S1820 — Ghost struct had 21 fields, exceeding the 20-field limit.
Merged flash_col and flash_row into a single flash_tile (encoded as
row * MAP_COLS + col), reducing the struct to 20 fields. Updated
ghost_respawn, ghosts_draw, ghosts_init_level, and tests accordingly.

+++ src/fruit.c
@@ -78,7 +78,7 @@ void fruit_draw(const Fruit *f, int offset_x, int offset_y) {
     if (f->popup_timer > 0.0f) {
         float progress  = 1.0f - (f->popup_timer / 0.8f);
         int   fx = (int)((float)f->popup_col * TILE_SIZE + TILE_SIZE / 2.0f) + offset_x - 12;
-        float fy = (float)f->popup_row * TILE_SIZE + offset_y - 4.0f
+        float fy = (float)f->popup_row * (float)TILE_SIZE + (float)offset_y - 4.0f
                    - progress * (float)TILE_SIZE * 1.5f;
         unsigned char alpha = (unsigned char)(255.0f * (1.0f - progress));
         DrawText(TextFormat("+%d", f->score), fx, (int)fy, 14, (Color){255, 255, 255, alpha});
